Recurrence rule performance improvements
Dynamically calculates appropriate start time for recurrence rules based on 1) the current time and 2) the time after which to get the next occurrence. This is done because the runtime of `rrule.after` grows linearly with the time delta between the rrule's `dtstart` and the date passed to `after()`, if `dtstart` is earlier (the rrule will calculate all the intermediate occurrences until the `after` date).